+++ src/arrFindValByProp.ts
@@ -1,6 +1,6 @@
 import { isObject } from './isObject'
 import { arrFindItemByProp } from './arrFindItemByProp'
-import type { PropKey, PropObj } from '../types'
+import type { PropObj } from '../types'
 
 export const arrFindValByProp: {
   /**
@@ -15,14 +15,14 @@ export const arrFindValByProp: {
    * @param propKey 属性键
    * @param propValue 属性值
    * @param rtnPropKey 返回属性
-   */<T extends PropObj>(array: T[], propKey: PropKey, propValue: unknown, rtnPropKey: string): unknown
-} = <T extends PropObj>(array: T[], propKey: PropObj | PropKey, propValue: unknown, rtnPropKey?: string): unknown => {
+   */<T extends PropObj>(array: T[], propKey:  PropertyKey, propValue: unknown, rtnPropKey: string): unknown
+} = <T extends PropObj>(array: T[], propKey: PropObj |  PropertyKey, propValue: unknown, rtnPropKey?: string): unknown => {
   let props: PropObj
   if (isObject(propKey)) {
     props = <PropObj>propKey
     rtnPropKey = <string>propValue
   } else {
-    props = { [<PropKey>propKey]: propValue }
+    props = { [< PropertyKey>propKey]: propValue }
   }
   const item = arrFindItemByProp(array, props)
   return item?.[rtnPropKey]